backend-python/requirements.txt
fastapi
uvicorn
httpx
psycopg2-binary
requests